+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Minor changes
 
 - Update npm packages and create package-loc.json from scratch, solving many remaining open audit hits
+- Update Ramda to 0.28.0, fix `merge` being removed in favour of `mergeRight` and `contains` in favour of `includes`
 
 ## Version 5.4.0